Some advances, still not working.

+++ src/types.rs
@@ -1,1 +1,1 @@
-pub type Particle = u8;
+pub type Particle = u32;
